Week 5. Iteration (while & do-while)
Write a program which takes keeps taking in numbers from a user on the command line until the user enters "quit" or "exit".  The user should be allowed to enter only one number each time and the system should compute the average and variance of all numbers entered so far each time.  To do this WITHOUT storing all the numbers we must use formulas which allow to compute the average and variance based only on the previous values (of the average and variance) together with the latest number entered.
AVERAGEN = AVERAGEN-1 + (NUMBERN - AVERAGEN-1) / N
VARIANCEN = ((VARIANCEN-1 * (N-1)) + (NUMBERN - AVERAGEN-1) * (NUMBERN - AVERAGEN)) / N