Week 8. Selection (switch) and Nested Loops
Write a program to produce the lyrics (all 12 verses) of the Twelve Days of Christmas. This is a cumulative song as each verse is build on top of the previous verses.  You must

have a loop for the verses,
build up each of the verses within a loop (or switch statement),
use at least one switch statement in your solution,
only write the lines of text (or constants representing them) from the song once in your program (e.g. you cannot have more than one statement in the entire program which prints out "my true love sent to me")...